requirements.txt
chromadb==0.6.3
dependency_injector==4.46.0
fastapi==0.115.12
google-api-python-client==2.164.0
groq==0.20.0
langchain
langchain_community==0.3.20
langchain_groq==0.3.1
langchain_openai==0.3.11
motor==3.7.0
numpy==2.2.4
protobuf==6.30.2
pydantic==2.11.1
pymongo==4.11.3
python-dotenv==1.1.0
sentence_transformers==4.0.1
SQLAlchemy==2.0.40
tiktoken==0.9.0
uvicorn==0.34.0
pytest
pytest-asyncio
pylint
google-api-core
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos
pymupdf==1.25.4
faiss-cpu==1.10.0
fitz==0.0.1.dev2
PyPDF2==1.26.0
langchain-core
langchain-huggingface==0.1.2
langchain-text-splitters
python-multipart
beautifulsoup4==4.13.3
tools==0.1.9